Added Source ID Refresh button and removed ID computation from load
Previous versions of CellBlender computed the Source ID (SHA1) every
time the addon was loaded. In order to eliminate that startup
computation, the source ID is now computed by the make file and saved in
a file called cellblender_id.py. This is helpful because that file is
part of the source code repository and that file can be searched for a
given Source ID to find the correct repository version. For the GIT
repository, such a search would be done with this command:

  git log -S'4936e8f204dbd09dee3e6240e843' -- cellblender_id.py

The search will then show the GIT commit where that particular ID was
found inside of the file "cellblender_id.py". That will identify the
source code used for that build - assuming it was updated via the make
file before that source code was pushed to the repository.

Because this source ID is contained in a file created by the make file,
it is possible that it might not be updated when the source code is
changed and pushed to the repository without running make. In that case,
the Source ID read from cellblender_id.py will be from a previous build,
and that incorrect Source ID will be shown in the CellBlender panel. For
that reason, a button was added next to the Source ID which will
recompute the Source ID from the files installed in the addon folder.
This can be used to refresh the Source ID if there's ever a question
about which version of the code is currently running at any time. At the
current time, that button only changes what is displayed in the panel.
It does not currently rewrite the cellblender_id.py file.