Added the cellblender link to .gitignore file
In order to create a zip file from this directory with all files in the
zip file prefixed by "cellblender/", there is a soft link named
"cellblender" which points to this same directory.

The soft link is created with:

  ln -s . cellblender

to keep the GIT status command from reporting that link as an untracked
file, it has been added to the .gitignore file.