Changes to makefile and CellBlender ID System
Previous versions had potential inconsistencies between the
cellblender_info.cellblender_source_list and what was included in the
makefile. One of the biggest changes in this commit is the
centralization of the complete addon file list inside the
"cellblender_source_info.py" file. That file can be run from inside a
makefile to produce a list of source file names compatible with use in
that same makefile. Consequently, the cellblender_source_file_info.py
file should contain the relative file names of all files to be included
in the cellblender.zip file with the exception of "cellblender_id.py".
This list will be used by the makefile to generate an ID file
("cellblender_id.py") containing the current ID. This same list will
also be used by the makefile when it builds the zip file. And finally,
this same list will be used to re-compute the CellBlender ID which is
displayed to the user for identification.

To make this work properly, the project directory where the makefile is
run must contain a symbolic link named "cellblender" which points to
that same project directory (ln -s . cellblender). This is needed so
that the files added to the zip file will be in a "cellblender"
subdirectory. So the zip file will contain files with a "cellblender/"
prefix (like cellblender/data_model.py).

Other changes:

-Removed cellblender_info from __init__.py and moved to
cellblender_source_info.py.

-Removed identify_source_version from __init__.py and moved to
cellblender_source_info.py.

-Added code to show the original ID from cellblender_id.py when it
differs from the dynamically computed (refreshed) version.

-Changed the default plot paging option to "One Page, Multiple Plots".

-Used "os.sep" in place of "/" in
cellblender_source_info.cellblender_source_list.

-Removed extra test page from Java plotting plug-in.

-Modified makefile to use file list created by running: python
cellblender_source_info.py

-Modified makefiles to use simple "make" command more effectively (not
requiring a "make clean")