Only regenerate the cellblender_id.py file when it would change.
The previous version rewrote the cellblender_id.py file every time. This
caused the make file to retrigger anything depending on that file every
time it was run. The current version computes the source ID, but then
checks to see if it would be rewriting the exact same ID. If so, then no
write is done and the old file date remains (no need for make to
rebuild).