Bugfix: file permissions error when creating log
CellBlender was trying to create a logfile in a directory that the user
might not have the appropriate permissions. For example, in Windows, it
might try to create the log file in C:\Program Files\...

It now writes to a log file in a temporary directory that the user
should essentially always have access to regardless of the platform
(e.g. /tmp/ on Linux and e.g. C:\Users\username\AppData\Local\Temp on
Windows).