Special handling of .jar file needed because of time stamp
It appears that rebuilding the .jar file (even without changing or even
recompiling any of the java classes) creates a jar file with different
contents every time it is rebuilt. This appears to be the result of some
sort of time stamp stored in the .jar file itself.

This creates a problem for our makefile/versioning system since it will
produce a different signature every time the make clean; make sequence
is run.

To get around this problem, I've removed the .jar file from the source
list and replaced it with the .java source file (appropriate). But for
the convenience of the end user, the .jar file will still be added to
the .zip file by the makefile.