Fixed jar file problem and added molecule colors.
It turns out that rebuilding a Java jar file changes it even if the
source and object (class) files haven't changed. I'm guessing this is
due to some form of time stamp included in the .jar file. This caused
our checksumming code to report a different version even if the same
version of the source was used. In other words, every time I ran make, I
would get a different cellblender_id.py file! I first suspected that it
was an error due to checksumming the previous checksum file (which would
cause a similar result). This ended up being a "wild goose chase" since
the real problem was a change to the jar file after each rebuild.

This problem arose from the desire to have the java plotting work "right
our of the box" without having to recreate the .jar file from source. So
the precompiled .jar file was included in the cellblender.zip file
directly (without requiring recompilation by end users).

To get this same functionality, I included the .java source file to be
checksummed and added the .jar file as a separate file (not checksummed)
when the .zip file is created. This was also done for cellblender_id.py
which is another file generated by the make file. Note that because our
cellblender.zip file contains the cellblender directory itself, these two
files had to be prefixed with "cellblender/" so they would be in the proper
location in the .zip file. This caused considerable confusion for a while
because the files were not being prefixed in my early versions of the make
file and so they weren't over-writing the ones already in my addon
folder.

In addition to those changes, this version also includes a minor upgrade
to the Java plotter which now uses the molecule colors when that option
is selected.