Experimenting with alternate simulation controllers
This particular commit attempts to replicate the SimControl.java
functionality with an OpenGL equivalent. Note that this would have to be
compiled on each platform ... just as MCell does. So this might be more
of an MCell front end than a CellBlender component.

This particular version is starting from a simple OpenGL application
that prints to the screen.