Added Run buttons for Java and OpenGL simulation controllers.
The make file should build SimControl.jar and SimControl executable.
The blue buttons in the interface will run each of them.

The OpenGL version doesn't have the ability to kill the jobs yet.
The OpenGL version uses GLUT which may not be directly available within
Blender. However, most of the actual OpenGL commands appear to be mapped
to Python equivalents. So if we can figure out how to open the top level
windows (and do anything else provided by GLUT) we might be able to
include a SimControl functionality completely within Blender.

Note that these changes were committed to the SimControl branch as
experimental at this time.