Added SimControl window position controls
Defined " : " as the separator between window controls and the
subprocess command for both Java and OpenGL implementations.

Also defined x=# and y=# as position commands in both versions.

Modified the operator code to produce the x=# and y=# commands to
stagger the resulting simulation windows to make them easier to see
and control.