Completed the preliminary importing of Data Model geometry.
Mostly added the surface regions when rebuilding the properties.

Note that some of the code still uses the model object list rather than
the mcell "include" flags attached to each object. Also note that some
of the current status (hidden,selected,...) may not be preserved when
importing a model. Check for "TODO" flags in the code.