Changing code to use "include" flags rather than model object list.
The include flags on each object are the definitive markers for which
objects belong in an MCell model. The previous version had used the
model object list, but that list isn't always correct (depending on
whether it's been updated or not). This change uses the "include" flag
within each object's "mcell" field to make this determination.