Added some test code to show a dummy value for lr_bar.
This seemed to update properly for some unknown reason. It may be that
Blender is redrawing the entire screen every time something changes?

This was tested and found to work on Blender 2.71, 2.70, 2.69, 2.68, and
2.66.

While the value displayed is not correct (it's just a dummy), this
version was still committed since it was the version tested on those 5
different versions of Blender.