Implemented lr_bar calculation and display.
Note that it seems somewhat surprising that this code updates the lr_bar
label without requiring any update functions. It appears that the panel
is redrawn properly as it is. This was tested on older versions of
Blender and it appears to work on versions as far back as 2.66. The
mechanism behind this auto update remains a mystery at this time.