Added object location to the data model
Objects in Blender have a location and a mesh relative to that location.
If the location for an object is left at the origin, then the object's
mesh will be relative to the origin. This works fine, but it means that
the object manipulator (arrows, resize handles, etc) will be located at
the origin rather than at the object's center. Since this is not the
normal behavior for objects created in Blender, a location field was
added to the geometry for each object. The location is currently a list
of three doubles rather than a dictionary of x,y,z.