Completed initial import of materials from the data model.
Note that this may not be the final data model representation for
geometry and materials. This current version stores the
"geometrical_objects" in the data model in its own dictionary entry
(separate from the model_objects dictionary entry). This is somewhat
redundant since the model_objects dictionary is just a list of object
names and that same list of object names is duplicated in the
geometrical_objects object_list. For that reason, it might make sense to
combine them. On the other hand, the geometry is currently considered an
"extra" part of the data model. Also, we might want to store non-model
objects in the data model, so combining them might be confusing.