Finished adding the actual molecules panel to cellblender main panel.
All of the panels in the cellblender main panel have been dummies up to
this point. This version is the first to include an actual CellBlender
panel (molecules, in this case) in the main panel. Because the molecules
panel appears in both locations (main panel and molecules panel), any
changes made to either will appear in both.