Better separation of draw_panel and draw_layout.
The original molecule code used "draw_panel" which accepted a panel
"widget" in which to draw. But there may be times when we want to draw
into an existing layout (rather than a panel). So the previous draw
panel code was separated into draw_panel and draw_layout functions. This
change simplifies the draw_panel code so that almost all of the work is
done in the draw_layout function.