Removed global referecences in preferences / added init button
One of the benefits of moving toward an object-oriented programming
style is the ability to reference most member data via "self.item"
rather than "context.scene.mcell.class.item". This change is not
strictly necessary, but it simplifies the code, makes it more readable,
and helps highlight the locations where global references are really
needed (possibly suggesting better factoring of the code). This change
will eventually be repeated for many other property groups.

The CellBlender main panel was also set to default in the open state
(by removing 'DEFAULT_CLOSED'), and the initialize button was added.