Implemented molecule color selection within the molecules panel.
This allows selection of molecule colors inside the molecule panel. This
implementation includes a check box to enable use of the molecule color
over the material color. Changing the molecule color effectively changes
the material color. However, by disabling the check box, the color can
be completely manipulated via the materials panel as was previously
supported by CellBlender.

The implementation of molecule glyph and scale is not completed yet.