Consolidated Initialization and Run Panels, Added Run Control Selector
The Model Initialization and Run panels were merged so the number of
iterations and other run-related settings (output, notifications, etc)
would be immediately visible at the time of pushing the "Run" button.

Note that this change somewhat violated our object-oriented separation
since this particular drawing function needs to display data from both
the model initialization object and the run_simulation object. As a
result, the drawing code is a little more complicated than it was
before. This might suggest a refactoring of those two classes.

This commit also included a merging of the three former run buttons
(run, run java, and run opengl) into a single run button with an "enum"
selector to choose the actual run method.