Added option to show buttons with text in CellBlender Preferences
In accordance with yesterday's meeting, an option has been added to show
long menu buttons (buttons with text). This option is at the bottom of
the CellBlender Preferences panel (beside the buttons which show/hide
the older panels). When this option is selected, the button bar will be
replaced with two columns of buttons containing both icons and text.
These buttons are actually booleans (just like the button bar) and can
be operated in either exclusive mode or multiple mode according to the
push pin setting (just like the button bar).