First cut at merging molecule color selection / Console Autocomplete
Setting the molecule color can be tricky since the original CellBlender
created materials on the fly as molecules were read. So in order to
allow users to set molecule colors first, we either need to create the
materials as molecules are defined, or transfer the user colors to the
actual materials as created. Currently the latter approach is being
attempted.