Began testing current version against .blend files from CB1.RC3.
Unfortunately, it appears that having changed the default molecule
type from 2D (requiring a defined surface) to 3D (much simpler),
also caused an incompatibility in the .blend files such that files
from RC3 would generate MDL files which failed when run with MCell.
So for now, the default molecule type was returned to 2D.

This points out the kinds of unexpected problems we might encounter
as we try to maintain compatibility between .blend files created with
different versions of CellBlender. Hopefully, the data model will be
a good solution to these kinds of problems.