Duplicated the Defined Surface Regions panel in the new panel system.
Added a draw_layout function separate from draw_panel in
object_surface_regions.py so the code could draw to a panel or a layout.

Note that the Surface Region interface could use some upgrading. For
example, it would be good to select objects in the 3D window when they
are selected in the Model Objects window (and vice versa).