Added copying capability to Java Simulation Runner
For debugging, it was handy to be able to detect MCell errors from
normal MCell output. The Java Simulation Runner opens two streams
to the running MCell process (stdout and stderr) and displays them
differently in the window. However, it didn't provide a way to
select just one or the other (or even both) and copy them to the
system clipboard. This change provides that capability through the
"Copy Normal Output", "Copy Error Output", and "Copy All Output"
menu options.