Triangulate meshes when added to Model Objects.
Mesh objects will automatically be triangulated when added to the Model
Objects list. Since MCell requires meshes to be triangulated anyway, it
makes little sense to allow users to add meshes that aren't triangulated
to the Model Objects list. This removes one additional step that a user
needs to do to make a working model and is in line with recent efforts
to simplify the CellBlender user experience.