Changed makefile and source_info for upcoming release of Blender 2.73.
Note that the save_pre bug fix will be in Blender 2.73, and this is
needed for proper data model operation with the startup file. So for
this reason, I've changed the compatability list to only include 2.7.3.
Note that this will not prevent running with earlier versions, but it
will issue a warning.