Orthogonalization of CellBlender Panel Display Options
These changes support showing the main CellBlender button
panel in either the original scene tab or the new tool shelf
CellBlender tab or both. It also supports showing the old
individual CellBlender Panels in the scene tab - although
the upgrade option is not displayed in that older mode.

These changes were made in an attempt to isolate a very slow
upgrade from older .blend files to the current version. It did
not seem to make any difference.