Fixed checking for mod_surf_regions.
When importing a data model, not everything needed is present at each
stage of the import. For example, when importing the mod_surf_regions,
the objects themselves may not have been imported yet. To prevent a
possible "chicken or egg" situation, the checking is done after the
importing of all property groups with "check_properties_after_building".
This is typically a call to the "check" functions for each member of a
list, and typically a call to the older "check" function from within
each member of the list. Note that the older "check" functions were
based on the active property within the list, and the "self" passed in
was not used. This change split the actual checking into a function that
does check itself (as called when importing) and a function to be called
as an update callback on the active item. The active item version is
called check_active_mod_surf_regions, and it just calls the
"self-centered" version (check_mod_surf_regions) on the active item.
This follows the object-oriented model more closely.

Note that this commit only does the checking for mod_surf_regions. The
same procedure still needs to be implemented for other property groups
that need checking after building.