Added load_pre printout, added "Upgrade" button to Project Panel
A load_pre handler was added primarily for debugging to mark the
location in the output where the loading began. However, it seems that
this is not being called when the startup.blend file is loaded. This may
be a bug in Blender, but it's not important for CellBlender at this
time.

For some reason, when the CellBlender Panels are drawn with the new
button display - for some .blend files - Blender locks up for a long
time (tens of seconds). This hasn't been seen for newly created .blend
files, but appears to be related to upgrading older .blend files. When
this happens, the CPU usage is at 100% and the entire Blender
application becomes unresponsive (does not even refresh itself). In
order to work around this, the older "Scene" panels have been enabled as
an option (see earlier commit). However, those older Scene panels did
not have an "upgrade" button (it's been implemented in the newer panels
as an option that's only presented when the .blend file ID doesn't match
the addon ID). So to allow upgrading with the older panels, an "Upgrade"
button has been added to the Project panel.