Added cellblender.zip to git repository for general use.
There may be CellBlender users who want to try the latest version but
are not able or willing to install a full build library. So including
the .zip file makes the latest version available to more end users. Note
that we may want to purge older versions when we make official releases
just to save space in the repository.