Changed "NamePanelProperty" to "NamePropertyGroup"
This change reflects the fact that the property groups are not
necessarily associated with panels. The property groups are groups of
related properties that form an "object" of some sort. These objects may
include drawing methods, but their rendering in a "Panel" is secondary
to their function of storing the CellBlender data itself.