Updating Molecule Display Controls
This version updates the material color based on the molecule's own
color property (rather than the other way around). This may cause a
conflict when changing the colors via the materials panel since the
molecules panel color will over-ride it when any changes are made to any
of the molecule's display properties. This was done explicitly for the
upcoming MCell class so changing the molecule colors could be done
primarily through the CellBlender interface.

The ability to scale molecules was also added in the CellBlender
interface. The current version rescales the glyph by multiplying by the
current scale factor and dividing by the previous scale factor. This
might introduce some loss of precision over time, but it should be
negligible for most practical uses. To avoid division by zero, the scale
cannot go below the arbitrarily chosen value of 0.0001 set as the "min"
value for the "scale" property.