Molecule Glyph Settings working in Molecules Panel
This change basically duplicates the molecule glyph setting code in the
molecule display settings panel. Rather than using the currently
selected shape (required by the old panel), it uses the currently
selected molecule for the molecule settings panel (where it resides).

Note that this implementation still suffers from some of the limitations
of the previous approach, but the user interface is much easier to use
since it's all contained in the CellBlender tab.

Also note that there's still an error message that is displayed the
first time reading viz data, and there are lots of debug statements that
need to be removed. They were preserved in this commit before being
removed.