Added print statements in MPL_Simple plotter for debugging.
For some unknown reason, the plotting functions didn't work in the class
we were teaching today. It's likely that it was a computer configuration
problem, but it was hard to debug without knowing exactly what command
was being executed. So the MPL_Simple plotter was modified to print this
information to help debug the configuration problem.