Attempts to keep cellblender.zip from changing after make clean; make
For some reason, simply remaking cellblender.zip changes it. This causes
git to store a new copy of the file in the repository resulting in the
git repository growing much larger than it needs to be.

This commit also introduces mdl_to_data_model.py. This is a simplistic
start at an MDL importer which currently only imports geometry and also
makes a lot of assumptions about the format of the MDL file (to simplify
the parsing).