Used git rm to remove cellblender.zip from git.
cellblender.zip was added as a convenient way to distribute the latest
version of cellblender. However, running "make clean" and "make" seems
to create a different version of cellblender.zip every time (most likely
due to time stamps in the .zip file). So even if no source files were
changed, remaking would create a new cellblender.zip. This was not a
problem when working on a single branch, but switching between branches
requires a re-running of make and this changes cellblender.zip every
time. This would needlessly add many copies of the same .zip file with
only compilation dates differing. For that reason, cellblender.zip was
removed from this branch.