Attempts to provide for renaming of model objects in CellBlender
This commit includes a variety of attempts (mostly commented out) to
allow for renaming of model objects within the Model Objects Panel.

This was part of an ongoing effort to centralize most of the commonly
needed model building functionality within the CellBlender panels which
minimizes the need to learn the rest of the Blender interface.

In this case, the renaming of model objects could not be done within any
of the CellBlender panels. The typical alternate is to use the outliner.
For ease of learning, the ability to rename an object was added to the
Model objects panel.

This was attempted in many ways. The favored approach would allow the
renaming of an item when selected in the list. However, there are some
problems because the object's name and the name in the model objects
list are two different strings. A number of attempts were made to fix
this, and they are mostly commented out in this version because they
didn't work completely. As a compromise, the model objects panel now
shows the name of the active object which can be changed BEFORE the
object is added to the model objects list. If it's changed afterward, it
will break the relationship between the model objects list and Blender's
actual objects list (this has been an ongoing problem and is not
resolved in this commit).

The handling of model objects needs more work to operate seamlessly
within Blender. Ideally the selection in the list and selection in the
3D view should mirror each other. And renaming either the Blender object
or the "model" object should rename the other (ideally they should be
referencing the same exact property). So this change is really just a
quick way to allow renaming a model object without having to hunt around
for the outliner in Blender's interface. It contains a lot of commented
ideas on better approaches, so they were left in this version as hints
for future improvement.