Various changes trying to track down the slow-down problem.
It appears that upgrading an RC3 file by loading its data model produces
a different result than loading from an external data model. In the
former case, there appears to be a significant delay when loading mol
viz files. In the latter case, this delay doesn't seem to appear. Even
more interesting, the delay disappears when the new CellBlender button
panel isn't being displayed. Note that a .blend file with a delay can
sometimes be used to export a Data Model and then that Data Model can be
imported into a fresh .blend file to eliminate the delay problem.

This change includes some debugging code as well as some code attempting
to fix problems that might be related to the delay problem.