Reintroduced Simulation Control selector under Output/Control Options.
It turns out that the code for all simulation runners still works, but
the selection control had been removed to simplify the interface. By
placing the selection control in the Output/Control Options subpanel,
the ability to run alternate simulation runners is available to users
who want to use it. The default remains "Queue Control" for those who
don't want to use the alternates.