Bugfix: matplotlib plots not working in Windows.
This fixes a problem with the matplotlib plotters in Windows. CB
couldn't find python in the PATH, since it was looking for python and
not python.exe. Additionally, this fix simplifies the code in a platform
independent manner (using shutil.which).

Manually setting the python path still doesn't work and should be fixed.
Also, the other plotters (java, xmgrace) weren't updated.