Cleaned up registration / unregistration of panels a bit.
Added logic to ensure that all CellBlender panels couldn't be shut.

Recommented the old Scene Panel registration which had been used for
testing.

Unregistered the old run_simulation panel in the "register" function
because the module registration had registered it. In order to show
panels in the proper order, we had been unregistering and reregistering
them. With the addition of the simulation queue runner, the
unregistration of the old run_simulation panel had been removed which
caused it to show up since it was already registered via the module
level registration.