Use plain python list, not a collection property for mol_file_list
Apparently there is some kind of internal inefficiency in Blender when
collection properties get too large.

The mol_file_list can get to be many 10's of thousands in length.  When
implemented as a collection property this slows things way down when displaying
certain other properties in panels even though the mol_file_list itself
seems to perform at full speed.  Will report this to as a Blender bug when
we can make a simple case that replicates the issue.