Added draw_prop_with_help function to support help with non-parameters.
The parameter property group has a provision for storing and displaying
help for each "panel parameter". This is provided through the "info"
button that's drawn next to each panel parameter. However, in
CellBlender we draw other items (enums, booleans, strings, etc) which
are not panel parameters. It would be nice to have a help system that is
compatible with the panel parameter help system for consistency. The new
"draw_prop_with_help" function was added to the ParameterSystemPropertyGroup
to provide this functionality. The calling function must provide:

  a layout to draw the item
  a label (string) to draw beside the item
  the property group of the item
  the name (string) of the item
  the name (string) of the boolean help flag for the item
  the boolean help flag itself (we may be able to eliminate this)
  the help string itself with newlines ("\n") to separate lines

This means that each item needs an additional BoolProperty associated
with it to store the state of the help (showing or hidden). This boolean
property is bundled into the panel parameters, but must be managed
separately with items not inside the parameter system.

This commit includes the support functions and implementation of help
for some of the molecule items:

  Molecule Name
  Molecule Type
  Target Only flag