Added info for almost all Molecule Release / Placement Properties
Completed adding information buttons for all of the non-prop_search
properties in the Molecule Release/Placement fields. However, two of the
fields (Molecule name and Release Pattern name) are not normal
properties (String, Boolean, Enum) and are instead created by the
prop_search method of the layout. This may require a variation of the
draw_prop_with_help function.