Added try/except to catch mol_viz_update error and added helptext
For some reason, we continue to generate an exception when first reading
the Molecule Visualization data. The exception pops up a large dialog
under Blender. Dismissing it and just reloading seems to work fine. So
as a workaround for this release, I've added a try/catch block around
the molv_viz_update function in the execute operator for read_viz_data.
If an exception is caught it will print it to the console rather than
popping up the dialog box. We should try to figure out what's causing
this.