Disable run button while simulation are running
This avoids accidentally hitting the run button twice and deleting/corrupting
output.

This easy to implement for the queue sim_runner because it tracks the status of
the job queue.  But the other sim controllers don't keep track so they are not
yet compatible with enforcing this policy.  I have disabled the sim controller
drop-down selector until the other sim controllers can be made to comply with
the policy.

This means the queue sim runner is the default and only choice for now.