Clean up makefile, remove jar files from cellblender_source_info.py
  I made a new variable in the makefile called ZIPFILES so that SOURCES can be
just the list of sources needed to compute the cellblender_id and not all the
bits needed in the zip file.  Previously SOURCES included the jar files which
are built from the java files but really just the java files are sources.
Items that are built are not sources.

I also removed the jar files from the repo.  I'm hoping that if the list of
items in the SOURCES exactly matches what we have in the repo then the SHA-1
from SOURCES will exactly match the git hash.  This means all the makefiles
must also be listed in SOURCES.