Handling of Arbitrary General Parameter Order
CellBlender allows the General Parameters to be defined in any order.
So this is OK in CellBlender:

  a = 2 * b
  b = 3

The CellBlender parameter system will figure out that "a" is 6.

When parameters are normally exported for MCell, they are normally
evaluated and exported as constants producing this:

  a = 6
  b = 3

This solves the "forward reference" problem, and that's why we
haven't generally had problems in the past. However, there is a
flag in the Parameter Options panel listed as:

   Export Parameters as Expressions (experimental)

When that flag is set (checked), the MDL will contain the
expressions rather than the numeric values. In that case,
the order-dependency is not checked at export time, and
MCell will currently generate a run-time error like this:

  Fatal error: On line: 3 of file Scene.parameters.mdl
  Undefined variable: b
  Task PID 20439  status: mcell_error  return code: 1

So the order dependency issue has to be resolved somewhere
to fully fix this issue. It could be argued that it would
be better to fix it in MCell so all MDL could benefit from
defining variables in arbitrary order. However, it's also
something that could be done in CellBlender when the MDL
is written. I suspect the work to be done is similar in
either case, so that argues to do it in MCell. On the other
hand, the CellBlender Parameter System might have more code
to help with that process than MCell does.

For now, the work-around is to either define parameters
in the order that they would be defined in MDL or to set
the "Export as Expressions" flag as False (unchecked). That
is the current default, so this is only an issue for folks
that have changed it the Parameter Options panel.

Additionally, this issue uncovered a deficiency in reading the
data model parameters into Blender properties. As with MCell,
the code that read the parameters performed an evaluation of
each parameter based on what had appeared (been read) already.
This created error warnings in the CellBlender panel that would
go away by re-entering the same expression. To resolve this
problem, code was added to check the "isValid" status of each
parameter and update the invalid status repeatedly until either
all are valid or until the number of invalid parameters isn't
changing (Parameter System / build_properties_from_data_model).