Exporting the Value when an Expression is Blank
The previous commit mentioned a problem when exporting parameters
that use the convention that a blank expression implies the default
numeric value. As mentioned in that message, there are cases when
a calling function might need to get the raw string expression from
a parameter, and also cases when a calling function might want the
expression if non-blank and the numeric value (default) when the
expression is blank. This commit adds a function to do that.