Integrated proper rxn_output export to bng importer
previously a special class invisible to the user and the gui was used as a
proxy for this. Now  a special attribute to MCellReactionOutputProperty was
added to the data model and the gui that handles literal mdl output that the
user can input. This is used by the bng exporter for observable handling.

The old classes were commented out and left in for the time being for reference
purposes.