Copy of master with stashed changes for implementing "checkonly".
This commit is a copy of master that was being modified to include a
"checkonly" option for build_properties_from_data_model. This approach
became a problem because the build_properties_from_data_model function
needed to construct sub objects in order to call their own
build_properties_from_data_model functions. So the molecule list would
check its own properties and then create each molecule and call it's
check properties. But because properties are actually being created
(molecules in this example), this makes a bit of a mess that would need
to be cleaned up.

This problem is caused by including the "data model surgery" in the
build_properties_from_data_model code - that's actually building the
properties. It would be much better to do the data model updating first
(without changing any properties at all). Then if the resulting
updated data model is OK, the building of properties from the data model
could begin with no further checking required. More importantly, if the
data model could not be updated, then the properties would remain as
they had been (without having been partially changed in the update
process).

So if the "data model surgery" is done without building the actual
properties, then the result can be applied or discarded depending on
whether the resulting "upgraded" data model is consistent with the
current version of CellBlender.

Of course, the code that does the "surgery" on the data should be a
member of the class that's being operated on. This would require (in
Java terminology) a class method (or static method) since it would need
to be called without an actual instance of the object existing. So, for
example, the molecule class would need a function to upgrade its data
model without requiring an actual molecule instance (Blender Property
Group) to be created. In Python, this is accomplished with the
@staticmethod decorator (as opposed to the @classmethod decorator).

For that reason, the code in this commit will be dropped in favor of a
separate function that implements the data model upgrade process.